sync

+++ src/dMRI/tractography/io/tractogramReader.h
@@ -44,11 +44,11 @@ namespace NIBR
     };
 
     struct Segment {
-        int     streamlineNo;
-        double  p[3];
-        double  dir[3];
-        double  length;
-        void*   data;
+        int    streamlineNo;
+        float  p[3];
+        float  dir[3];
+        float  length;
+        void*  data;
     };
 
     typedef enum {
